Split cli into main/type/parser modules
